.history/requirements_20260321122534.txt
streamlit
pandas
numpy
yfinance
plotly
scikit-learn
xgboost
prophet
ta
scipy
# transformers
# torch
newsapi-python
python-dotenv